+++ src/servidor.js
@@ -43,6 +43,7 @@ function crearApp() {
   app.use('/api/existencia', require('./modulos/existencia/rutas'));
   app.use('/api/caja', require('./modulos/caja/rutas'));
   app.use('/api/ventas', require('./modulos/ventas/rutas'));
+  app.use('/api/ayuda', require('./modulos/ayuda/rutas'));
   app.use('/api/versiones', require('./modulos/versiones/rutas'));
   app.use('/api/sistema', require('./modulos/sistema/rutas'));